fixed_location_of_report
